requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.11.0
djangorestframework-simplejwt>=4.7.2
psycopg2-binary>=2.8.6